*: make config groups optional
Previously, both the `reporting` and `collecting` groups had to
be specified in every config snippet in order for the TOML
parsing to succeed. Instead make the groups and their keys
optional, and later perform validation for keys that are
required (such as `reporting.enabled`) once all the snippets
have been parsed.

This implicitly makes the default `collecting.level` value
`"minimal"`, which only takes effect if metrics reporting is
enabled.