Use DynamicUser=yes
I was looking at `/etc/passwd` on FCOS and noticed a user allocated
for this service.

With modern systemd, I think `DynamicUser=yes` is exactly what
we want for this type of service.  We don't need to allocate
a uid/gid persistently on the system.  It's just a lot cleaner
this way.

Drop the `tmpfiles.d` snippet too; this means the admin would need
to `mkdir` that directory.  Which is probably better actually,
because we don't want the directory to actually be *owned* by the
service user since then it'd be mutable by the service.  The
admin just needs to make the directories world executable and
the files world-readable.