service: drop requirement on network-online.target
As per systemd recommendations[1], we should avoid depending on this
unit. Instead, if/when we do implement the reporting bits, we should be
resilient to networks taking time to come online or temporary network
flakes.

This is the only thing which is actively pulling in
`network-online.target` right now (the other one being `docker.service`,
but we have less control over that one, and it's disabled by default).

[1] https://www.freedesktop.org/wiki/Software/systemd/NetworkTarget/